side-bar-submenu fix
