+++ angular-src/src/app/components/dashboard/admin/dash-project-detail/dash-project-home.component.ts
@@ -218,9 +218,7 @@ export class DashProjectHomeComponent implements OnInit {
             break;
           }
         }
-
       }
-
 
       if (this.canManageProjects) {
         this.users = allusers;
