+++ angular-src/src/app/components/dashboard/client/floorplan/dash-floorplan.component.js
