+++ angular-src/src/app/components/dashboard/client/floorplan/floorplan.component.html
