+++ angular-src/src/app/components/dashboard/dashboard-routing.module.ts
@@ -3,14 +3,12 @@ import {DashboardComponent} from "./dashboard.component";
 import {AuthGuard} from "../../guards/auth.guard";
 import {DashUsersComponent} from "./admin/dash-users/dash-users.component";
 import {NgModule} from "@angular/core";
-import {UserProvider} from "../../services/user.service";
 import {DashOverviewComponent} from "./client/project-overview/dash-overview.component";
 import {DashProjectsComponent} from "./admin/dash-projects/dash-projects.component";
-import {ProjectProvider} from "../../services/project.service";
-import {ProfileComponent} from "./client/profile/profile.component";
 import {DashProjectHomeComponent} from "./admin/dash-project-detail/dash-project-home.component";
 import {DashFloorplanComponent} from "./admin/dash-floorplan/dash-floorplan.component";
-import {FloorplanProvider} from "../../services/floorplan.service";
+import {ClientFloorplanComponent} from "./client/floorplan/floorplan.component";
+import {ProfileComponent} from "./client/profile/profile.component";
 
 const dashRoutes: Routes = [
 
@@ -24,7 +22,14 @@ const dashRoutes: Routes = [
       {path: 'projects/:projectid/floorplan/:floorplanid', component: DashFloorplanComponent},
 
 
-      {path: 'project/:id', component: DashOverviewComponent}
+      {path: 'project/:id',  children: [
+          {path: '', redirectTo: 'home', pathMatch: 'full'},
+          {path: 'home', component: DashOverviewComponent},
+          {path: 'floorplan/:floorplanId', component: ClientFloorplanComponent},
+          {path: 'contents/:id', component: DashUsersComponent}
+
+        ]},
+
     ]},
 ];
