+++ angular-src/src/app/components/dashboard/sidebar/sidebar.component.html
@@ -21,11 +21,11 @@
 
 <ul class="nav-list nav nav-pills flex-column">
   <li class="nav-item" *ngFor="let project of projects" >
-    <a class="nav-link" [routerLink]="'project/'+project._id" [routerLinkActive] = "['active']">{{project.name}}</a>
+    <a class="nav-link" [routerLink]="'project/'+project._id + ''" [routerLinkActive] = "['active']">{{project.name}}</a>
 
     <ul [routerLinkActive] = "['submenu']" class="nav-list">
       <li class="nav-item">
-        <a class="nav-link" [routerLink]="'project/'+project._id" [routerLinkActive] = "['active2']">Home</a>
+        <a class="nav-link" [routerLink]="'project/' + project._id +'/home' " [routerLinkActive] = "['active2']">Overview</a>
        </li>
 
       <li class="nav-item" *ngFor="let floorPlan of project.floorPlans; let i = index">
@@ -33,7 +33,7 @@
       </li>
 
       <li class="nav-item">
-        <a class="nav-link" [routerLink]="'contents/'+project._id" [routerLinkActive] = "['active2']">Contents</a>
+        <a class="nav-link" [routerLink]="'project/'+project._id+'/contents/'+project._id" [routerLinkActive] = "['active2']">Contents</a>
       </li>
     </ul>
